+++ src/playlistcontainer.c
@@ -278,6 +278,37 @@ PlaylistContainer_sq_length(PyObject *o)
     return sp_playlistcontainer_num_playlists(self->_playlistcontainer);
 }
 
+static PyObject *
+PlaylistContainer_add_new_playlist(PlaylistContainer *pc, PyObject *args)
+{
+    char *name = NULL;
+    int len;
+    sp_playlist *playlist;
+
+    if (!sp_playlistcontainer_is_loaded(pc->_playlistcontainer)) {
+        PyErr_SetString(SpotifyError, "PlaylistContainer not loaded");
+        return NULL;
+    }
+    if (!PyArg_ParseTuple(args, "es#", ENCODING, &name, &len))
+        return NULL;
+    if (len > 255) {
+        PyErr_SetString(PyExc_ValueError,
+                "Playlist name must be < 255 characters long");
+        goto error;
+    }
+    playlist =
+        sp_playlistcontainer_add_new_playlist(pc->_playlistcontainer, name);
+    if (!playlist) {
+        PyErr_SetString(SpotifyError, "Operation failed.");
+        goto error;
+    }
+    return Playlist_FromSpotify(playlist);
+
+    error:
+        PyMem_Free(name);
+        return NULL;
+}
+
 /// PlaylistContainer Get Item []
 PyObject *
 PlaylistContainer_sq_item(PyObject *o, Py_ssize_t index)
@@ -331,6 +362,10 @@ static PyMethodDef PlaylistContainer_methods[] = {
      (PyCFunction)PlaylistContainer_add_playlist_removed_callback,
      METH_VARARGS,
      ""},
+    {"add_new_playlist",
+     (PyCFunction)PlaylistContainer_add_new_playlist,
+     METH_VARARGS,
+     "Add a new empty playlist to the playlist container."},
     {NULL}
 };